requirements.txt
httplib2
jsonschema
pyserial
requests
unittest2
urllib3
